refactor(languages): make language wave producers direct modules
