+++ crates/compass-languages/src/evidence/mod.rs
@@ -1,10 +1,14 @@
 mod build;
 mod csharp;
-mod extended;
+mod dart;
+mod groovy;
 mod kotlin;
 mod model;
 mod php;
 mod ruby;
+mod scala;
+mod shared;
+mod swift;
 mod typescript;
 mod validate;
