+++ crates/compass-languages/src/evidence/scala.rs
@@ -4,9 +4,9 @@ use std::path::Path;
 
 use tree_sitter::Node;
 
-use super::super::model::SemanticEvidenceBatch;
-use super::super::validate::EvidenceError;
-use super::common::{self, LanguageProfile};
+use super::model::SemanticEvidenceBatch;
+use super::shared::{self, LanguageProfile};
+use super::validate::EvidenceError;
 
 struct Scala;
 
@@ -14,12 +14,12 @@ impl LanguageProfile for Scala {
     const LANGUAGE: &'static str = "scala";
 
     fn package_name(source: &[u8]) -> Option<String> {
-        common::package_name_from_source(source)
+        shared::package_name_from_source(source)
     }
 
     fn declaration_kind(kind: &str) -> Option<&'static str> {
         let lower = kind.to_ascii_lowercase();
-        common::shared_declaration_kind(kind)
+        shared::shared_declaration_kind(kind)
             .or_else(|| (lower.contains("val_") || lower.contains("var_")).then_some("field"))
     }
 }
@@ -30,5 +30,5 @@ pub(super) fn emit_tree_evidence(
     source: &[u8],
     root: Node<'_>,
 ) -> Result<SemanticEvidenceBatch, EvidenceError> {
-    common::emit_tree_evidence::<Scala>(path, source_file, source, root)
+    shared::emit_tree_evidence::<Scala>(path, source_file, source, root)
 }
